Move HTML-meta extraction into rotero-translate (port step 1)
Phase 1 step 1 of the pure-Rust Zotero translator port: hoist the shared
"Embedded Metadata"-style extraction (meta tags + JSON-LD) out of the
connector's scrape.rs into a new rotero-translate::html_meta module so the
upcoming native translator registry can reuse it without a circular dep.

- rotero-translate: add scraper dep; new html_meta.rs with extract_from_html
  and the meta/JSON-LD helpers; re-export extract_from_html.
- rotero-connector/scrape.rs: reduced to the scrape_url fetch wrapper,
  delegating extraction to rotero_translate::extract_from_html. Public
  signature unchanged.

Pure refactor, no behavior change. Dep graph stays acyclic
(rotero-connector -> rotero-translate -> rotero-models).
