+++ crates/rotero-translate/Cargo.toml
@@ -13,3 +13,4 @@ tracing.workspace = true
 
 reqwest = { version = "0.12", features = ["json", "rustls-tls"] }
 directories = "6"
+scraper = "0.22"
